requirements.txt
# 📱 상품권 관리 시스템 - 안드로이드 앱 빌드 요구사항

# Kivy 및 KivyMD
kivymd==1.1.1
kivy==2.1.0

# 데이터베이스
sqlite3

# HTTP 클라이언트
requests==2.31.0

# 스케줄러
APScheduler==3.10.4

# 환경 변수
python-dotenv==1.0.0

# 이미지 처리
Pillow==10.0.0

# 암호화
bcrypt==4.0.1

# Base64 인코딩
pybase64==1.3.2

# SMS 발송 (선택사항)
twilio==8.9.1

# 안드로이드 빌드를 위한 Buildozer
buildozer==1.5.0

# Cython (컴파일 최적화)
Cython==3.0.2

# 네트워크 관련
urllib3==2.0.4

# JSON 처리
simplejson==3.19.1

# 시간 관련
python-dateutil==2.8.2

# 로깅
colorlog==6.7.0

# 파일 시스템
watchdog==3.0.0

# 푸시 알림 (선택사항)
pyfcm==1.5.4

# 백그라운드 작업
plyer==2.1.0
